requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
google-generativeai
langchain
langchain-google-genai
faiss-cpu